% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Title}
\usage{
birthday(x)
}
\arguments{
\item{x}{an integer that is the size of the group to sample}
}
\value{
a number which is the probability of at least one shared birthday
}
\description{
Title
}
\examples{
birthday(10)
birthday(40)
}
