% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance from the fire station (in miles).}
  \item{DAMAGE}{Estimated fire damage (in thousands of dollars).}
}
}
\source{
FIREDAM.csv provided in the MATH4753 course materials.
}
\usage{
fire
}
\description{
A dataset containing observations of distance from the fire station
and the resulting amount of fire damage.
}
\details{
This dataset is used in Lab 7 to explore relationships between distance
and damage in fire incidents.
}
\keyword{datasets}
