% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirst.R
\name{myfirstfunction}
\alias{myfirstfunction}
\title{Square a number}
\usage{
myfirstfunction(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector containing the square of x.
}
\description{
Takes a vector, squares it, and returns the result.
}
\examples{
myfirstfunction(10)
}
