% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets - Calculates how many tickets to sell before overbooking risk is greater then gamma (or 'pain' tolance of the customers)}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on the plane}

\item{gamma}{Probability of overbooking that is deemed OK}

\item{p}{Probability that a passenger shows up the day of the flight}
}
\value{
A list that has nd(discrete solution) and nc (normal approx), N, p, and gamma
}
\description{
ntickets - Calculates how many tickets to sell before overbooking risk is greater then gamma (or 'pain' tolance of the customers)
}
\examples{
ntickets(N=400, gamma=0.02, p=0.95)
}
