% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySample.R
\name{mysample}
\alias{mysample}
\title{mySample draws 'n' values with replacement from '1:10' and plots the relative frequency as a barplot.
Repeats this 'iter' times with pause between sample polling}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{integer. Sample size for each iteration}

\item{iter}{integer. Number of iterations (or the number of plots to produce)}

\item{time}{numeric. Seconds to pause between sample plots}
}
\value{
'NULL' due to divsion error
}
\description{
mySample draws 'n' values with replacement from '1:10' and plots the relative frequency as a barplot.
Repeats this 'iter' times with pause between sample polling
}
\examples{
mysample(n = 20, iter = 1, time = 1)
}
